requirements.txt
mysql-connector-python==8.2.0
bcrypt==4.1.2
pillow==10.1.0
qrcode[pil]==7.4.2
reportlab==4.0.7
pandas==2.1.4
matplotlib==3.8.2
openpyxl==3.1.2
pyserial==3.5
python-dateutil==2.8.2
